+++ src/utils/menu.py
@@ -441,7 +441,7 @@ injection.add_option("--prefix",
 
 injection.add_option("--technique", 
                 action="store",
-                default="cetf",
+                default="",
                 dest="tech",
                 help="Specify injection technique(s) to use.")